+++ shared/src/commonMain/kotlin/de/kitshn/ui/view/settings/SettingsServerAdvanced.kt
@@ -7,19 +7,29 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.Delete
 import androidx.compose.material.icons.rounded.Refresh
+import androidx.compose.material.icons.rounded.Security
 import androidx.compose.material.icons.rounded.Timer
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
+import androidx.compose.material3.Icon
 import androidx.compose.material3.ListItemDefaults
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.unit.dp
@@ -30,7 +40,11 @@ import de.kitshn.ui.component.settings.SettingsListItemPosition
 import de.kitshn.ui.dialog.TimeoutSelectionBottomSheet
 import de.kitshn.ui.dialog.rememberTimeoutSelectionBottomSheetState
 import de.kitshn.ui.view.ViewParameters
+import de.kitshn.utils.ClientCertificateData
+import de.kitshn.utils.rememberClientCertificateSelector
 import kitshn.shared.generated.resources.Res
+import kitshn.shared.generated.resources.action_abort
+import kitshn.shared.generated.resources.action_remove
 import kitshn.shared.generated.resources.settings_section_server_advanced_label
 import kitshn.shared.generated.resources.settings_section_server_advanced_long_timeout_description
 import kitshn.shared.generated.resources.settings_section_server_advanced_long_timeout_label
@@ -38,6 +52,12 @@ import kitshn.shared.generated.resources.settings_section_server_advanced_reset_
 import kitshn.shared.generated.resources.settings_section_server_advanced_reset_label
 import kitshn.shared.generated.resources.settings_section_server_advanced_short_timeout_description
 import kitshn.shared.generated.resources.settings_section_server_advanced_short_timeout_label
+import kitshn.shared.generated.resources.settings_section_server_mtls_description_none
+import kitshn.shared.generated.resources.settings_section_server_mtls_description_pkcs12
+import kitshn.shared.generated.resources.settings_section_server_mtls_label
+import kitshn.shared.generated.resources.settings_section_server_mtls_remove
+import kitshn.shared.generated.resources.settings_section_server_mtls_remove_dialog_description
+import kitshn.shared.generated.resources.settings_section_server_mtls_remove_dialog_title
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.stringResource
 
@@ -57,6 +77,36 @@ internal fun SettingsServerAdvancedContent(
     val longOptions = listOf(30000L, 60000L, 120000L, 300000L, 600000L)
 
     val timeoutSelectionBottomSheetState = rememberTimeoutSelectionBottomSheetState()
+
+    // mTLS state
+    val credentials = p.vm.tandoorClient?.credentials
+    val mtlsCertificateAlias = credentials?.mtlsCertificateAlias
+    val mtlsCertificateData = credentials?.mtlsCertificateData
+    val isMtlsConfigured = mtlsCertificateAlias != null || mtlsCertificateData != null
+    var showMtlsRemoveDialog by remember { mutableStateOf(false) }
+    val certificateSelector = rememberClientCertificateSelector()
+
+    fun applyMtlsCert(data: ClientCertificateData) {
+        val creds = p.vm.tandoorClient?.credentials ?: return
+        p.vm.updateCredentials(
+            creds.copy(
+                mtlsCertificateAlias = data.alias,
+                mtlsCertificateData = data.pkcs12DataBase64,
+                mtlsCertificatePassword = data.pkcs12Password,
+            )
+        )
+    }
+
+    fun removeMtlsCert() {
+        val creds = p.vm.tandoorClient?.credentials ?: return
+        p.vm.updateCredentials(
+            creds.copy(
+                mtlsCertificateAlias = null,
+                mtlsCertificateData = null,
+                mtlsCertificatePassword = null,
+            )
+        )
+    }
 
     LazyColumn(
         modifier = modifier,
@@ -127,7 +177,7 @@ internal fun SettingsServerAdvancedContent(
         }
 
         item {
-            Spacer(Modifier.height(16.dp))
+            Spacer(Modifier.height(8.dp))
         }
 
         item {
@@ -143,10 +193,67 @@ internal fun SettingsServerAdvancedContent(
                 }
             }
         }
+
+        item {
+            Spacer(Modifier.height(24.dp))
+        }
+
+        item {
+            SettingsListItem(
+                position = if(isMtlsConfigured) SettingsListItemPosition.TOP else SettingsListItemPosition.SINGULAR,
+                label = { Text(stringResource(Res.string.settings_section_server_mtls_label)) },
+                description = {
+                    Text(
+                        when {
+                            mtlsCertificateAlias != null -> mtlsCertificateAlias
+                            mtlsCertificateData != null -> stringResource(Res.string.settings_section_server_mtls_description_pkcs12)
+                            else -> stringResource(Res.string.settings_section_server_mtls_description_none)
+                        }
+                    )
+                },
+                icon = Icons.Rounded.Security,
+                contentDescription = stringResource(Res.string.settings_section_server_mtls_label)
+            ) {
+                certificateSelector.selectCertificate { data ->
+                    if(data != null) applyMtlsCert(data)
+                }
+            }
+        }
+
+        if(isMtlsConfigured) {
+            item {
+                SettingsListItem(
+                    position = SettingsListItemPosition.BOTTOM,
+                    label = { Text(stringResource(Res.string.settings_section_server_mtls_remove)) },
+                    icon = Icons.Rounded.Delete,
+                    contentDescription = stringResource(Res.string.settings_section_server_mtls_remove)
+                ) {
+                    showMtlsRemoveDialog = true
+                }
+            }
+        }
     }
 
     TimeoutSelectionBottomSheet(
         state = timeoutSelectionBottomSheetState
+    )
+
+    if(showMtlsRemoveDialog) AlertDialog(
+        onDismissRequest = { showMtlsRemoveDialog = false },
+        icon = { Icon(Icons.Rounded.Delete, null) },
+        title = { Text(stringResource(Res.string.settings_section_server_mtls_remove_dialog_title)) },
+        text = { Text(stringResource(Res.string.settings_section_server_mtls_remove_dialog_description)) },
+        confirmButton = {
+            Button(onClick = {
+                showMtlsRemoveDialog = false
+                removeMtlsCert()
+            }) { Text(stringResource(Res.string.action_remove)) }
+        },
+        dismissButton = {
+            TextButton(onClick = {
+                showMtlsRemoveDialog = false
+            }) { Text(stringResource(Res.string.action_abort)) }
+        }
     )
 }
